requirements.txt
biopython>=1.79
xlsxwriter>=3.0.0
